fuel_simulation/requirements.txt
Flask==2.3.2
grpcio==1.56.2
grpcio-tools==1.56.2